Crate fixed_size_lru_map

A fixed size map that keeps only the most recently used values. This can be used like a cache and it is thread safe.

Example

use fixed_size_lru_map::FixedSizeLruMap;

fn main() {
    let map = FixedSizeLruMap::with_capacity(2);
    let a = *map.get_or_init("a", || 10);
    let b = *map.get_or_init("a", || 12);

    assert_eq!(10, a);
    assert_eq!(10, b);
    assert_eq!(1, map.len());
}

Structs

FixedSizeLruMap

MapGuard